fix: footer issue and vaildtion schema
