+++ apps/galaxy/src/constants/index.ts
@@ -1,2 +1,1 @@
 export * from "./geolocation";
-export * from "./regex";
