--- apps/galaxy/src/constants/regex.ts
@@ -1,6 +1,0 @@
-export const phoneNumberRegex =
-  /^(?:\+?(61))? ?(?:\((?=.*\)))?(0?[2-57-8])\)? ?(\d\d(?:[- ](?=\d{3})|(?!\d\d[- ]?\d[- ]))\d\d[- ]?\d[- ]?\d{3})$/gm;
-
-export const postcodeRegex = /[0-9]{4}/;
-
-export const abnRegex = /[0-9]{11}/;
