+++ apps/galaxy/src/modules/auth/SignUpForm/RegisterInfoPage.tsx
@@ -14,7 +14,7 @@ const RegisterInfoPage = (props: {
   const { formStep, setFormStep, data, setData } = props;
   const { username, password, confirmPassword } = formConfig;
   const { register, formState, handleSubmit } = useForm<RegisterInfoFormData>({
-    mode: "onBlur",
+    mode: "all",
     reValidateMode: "onChange",
     resolver: yupResolver(RegisterInfoFormSchema),
   });
