+++ apps/galaxy/src/pages/_app.tsx
@@ -24,9 +24,9 @@ const App = ({ Component, pageProps }: AppProps) => {
       });
   };
 
-  useEffect(() => {
-    checkTokenExpiration(token) && tokenExpired(token);
-  }, []);
+  // useEffect(() => {
+  //   checkTokenExpiration(token) && tokenExpired(token);
+  // }, []);
 
   return (
     <ThemeProvider theme={randomTheme}>
