+++ package.json
@@ -32,5 +32,8 @@
   "engines": {
     "node": ">=14.0.0"
   },
-  "packageManager": "npm@8.5.0"
+  "packageManager": "npm@8.5.0",
+  "dependencies": {
+    "yup-phone-lite": "^1.1.2"
+  }
 }
